requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
PyMuPDF==1.23.8
langchain==0.0.350
langchain-openai==0.0.2
langchain-pinecone==0.0.1
langchain-core==0.1.0
python-dotenv==1.0.0
pydantic==2.5.0
